+++ outputs.tf
@@ -60,6 +60,16 @@ output "security_group_arn" {
   value       = element(concat(aws_security_group.this.*.arn, list("")), 0)
 }
 
+output "worker_security_group_id" {
+  description = "ID of the security group that is created for the workers"
+  value       = element(concat(aws_security_group.worker.*.id, list("")), 0)
+}
+
+output "worker_security_group_arn" {
+  description = "ARN of the security group that is created for the workers."
+  value       = element(concat(aws_security_group.worker.*.arn, list("")), 0)
+}
+
 output "kubernetes_version" {
   description = "Version of the EKS cluster."
   value       = element(concat(aws_eks_cluster.this.*.version, list("")), 0)